Kernel/GPU: Rename VirtIOGPU directory => VirtIO
The new-old subsystem is already called the GPU subsystem, so there's no
need for this kind of redundancy.
